rename deploy -> upload
